general cleanup and fewer unwraps
